IK-115 Refactor the location of Matcher class
Matcher class has now been moved to DomainModel folder
